requirements.txt
streamlit>=1.36
sqlalchemy>=2.0
PyMySQL>=1.1
python-dotenv>=1.0
pandas>=2.2
scikit-learn>=1.3
